requirements.txt
wavelink
psutil
humanize
aiohttp
feedparser
discord.py[voice]
PyYAML
pandas